scripts/ladderlog_uploader_test/requirements.txt
appdirs==1.4.3
cachetools==3.1.1
certifi==2020.6.20
chardet==3.0.4
configparser==4.0.2
contextlib2==0.6.0.post1
crcmod==1.7
distlib==0.3.0
filelock==3.0.12
futures==3.3.0
google-api-core==1.22.4
google-auth==1.22.1
google-cloud-core==1.4.3
google-cloud-storage==1.31.2
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
idna==2.10
importlib-metadata==1.6.0
importlib-resources==1.4.0
pathlib2==2.3.5
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.24.0
rsa==4.5
scandir==1.10.0
singledispatch==3.4.0.3
six==1.14.0
typing==3.7.4.1
urllib3==1.25.10
virtualenv==20.0.18
zipp==1.2.0